Revert "Alvos hitscan"
