+++ src/Entrega 1/Frontend/Monkey/Assets/Scenes/GameOver.unity
@@ -1158,6 +1158,7 @@ GameObject:
   - component: {fileID: 1488886045}
   - component: {fileID: 1488886044}
   - component: {fileID: 1488886043}
+  - component: {fileID: 1488886046}
   m_Layer: 0
   m_Name: EventSystem
   m_TagString: Untagged
@@ -1226,6 +1227,18 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1488886046
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1488886042}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 786f61a97ca90e74b8ea097322be27f3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1873549267
 GameObject:
   m_ObjectHideFlags: 0
