+++ src/Entrega 1/Frontend/Monkey/Assets/prefabs/alvo .prefab
@@ -13,8 +13,8 @@ GameObject:
   - component: {fileID: 1946333087136362715}
   - component: {fileID: 7313293556836084114}
   - component: {fileID: -4849135323334442568}
+  - component: {fileID: -3570578970324254250}
   - component: {fileID: 4748429172908317468}
-  - component: {fileID: 8114548904346334056}
   m_Layer: 0
   m_Name: 'alvo '
   m_TagString: Untagged
@@ -138,6 +138,19 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &-3570578970324254250
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 627160141094912426}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4920302949ffb324da499f24296097f2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  qtdAlvos: 3
 --- !u!135 &4748429172908317468
 SphereCollider:
   m_ObjectHideFlags: 0
@@ -159,15 +172,3 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.75
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &8114548904346334056
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 627160141094912426}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 77bac1df3802a6c499173a7f425cc52a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
