+++ src/Entrega 1/Frontend/Monkey/Assets/prefabs/alvoBonus.prefab
@@ -12,8 +12,8 @@ GameObject:
   - component: {fileID: 4269611418377219369}
   - component: {fileID: 28134040413160463}
   - component: {fileID: 195256085905780538}
+  - component: {fileID: 3983350375864681172}
   - component: {fileID: 5101691270621131442}
-  - component: {fileID: 797995701069058434}
   m_Layer: 0
   m_Name: alvoBonus
   m_TagString: alvoBonus
@@ -110,6 +110,20 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &3983350375864681172
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1275474290889906016}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2dbb428a158ef984ab7deb243f2eef9e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  textoAlvosBonus: {fileID: 0}
+  scriptPlayer: {fileID: 0}
 --- !u!54 &5101691270621131442
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -137,15 +151,3 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &797995701069058434
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1275474290889906016}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 77bac1df3802a6c499173a7f425cc52a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
