+++ src/Entrega 1/Frontend/Monkey/Assets/scripts/SpawnerControler.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-using UnityEngine.UI;
 
 public class spawnerControler : MonoBehaviour
 {
@@ -10,18 +9,6 @@ public class spawnerControler : MonoBehaviour
     public float tempoVida = 3f;
     public GameObject player;
     [SerializeField] Municao municaoScript;
-
-    Text textoAlvos;
-    public Text textoPontos;
-    public static int alvosAcertados = 0;
-    public static int qtdAlvos = 3;
-
-    private void Start()
-    {
-        alvosAcertados = 0;
-        textoAlvos = GameObject.FindWithTag("textoAlvos").GetComponent<Text>();
-    }
-
 
     void Update()
     {
@@ -64,11 +51,6 @@ public class spawnerControler : MonoBehaviour
                 movimento.InicializarSeguir(hit.collider.transform, veloMovimento, tempoVida);
 
                 Debug.Log("TIRO SEGUE: " + hit.collider.name);
-
-                alvosAcertados++;
-                textoAlvos.text = "Alvos:" + alvosAcertados;
-                playerMove.pontos += 5;
-                textoPontos.text = "Pontos:" + playerMove.pontos;
             }
             //verifica se acerto o trigger do alvo bonus
             else if (hit.collider.CompareTag("alvoBonus"))
@@ -88,30 +70,6 @@ public class spawnerControler : MonoBehaviour
                 movimento.InicializarSeguir(hit.collider.transform, veloMovimento, tempoVida);
 
                 Debug.Log("TIRO SEGUE: " + hit.collider.name);
-
-                playerMove.pontos += 10;
-                textoPontos.text = "Pontos:" + playerMove.pontos;
-            }
-            else if (hit.collider.CompareTag("inimigos"))
-            {
-                // Cria o tiro
-                GameObject novoTiro = Instantiate(tiro, posicaoPlayer, Quaternion.identity);
-
-                //pega o componente do movimento direcionado do tiro no outro scrpit
-                moviDirecionado movimento = novoTiro.GetComponent<moviDirecionado>();
-
-                if (movimento == null)
-                {
-                    movimento = novoTiro.AddComponent<moviDirecionado>();
-                }
-
-                // Inicializa com o movimento da funcao InicializarSeguir
-                movimento.InicializarSeguir(hit.collider.transform, veloMovimento, tempoVida);
-
-                Debug.Log("TIRO SEGUE: " + hit.collider.name);
-
-                playerMove.pontos += 15;
-                textoPontos.text = "Pontos:" + playerMove.pontos;
             }
             else
             {
