+++ src/Entrega 1/Frontend/Monkey/Assets/scripts/alvosBonus.cs
@@ -1,0 +1,28 @@
+using UnityEngine;
+using UnityEngine.UI;
+public class alvosBonus : MonoBehaviour
+{
+    public Text textoAlvosBonus;
+
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+    private void OnTriggerEnter(Collider hit)
+    {
+        if (hit.gameObject.tag == "projetil")
+        {
+            Destroy(gameObject);
+            playerMove.pontos += 10;
+            textoAlvosBonus.text = "Pontos:" + playerMove.pontos;
+            Destroy(hit.gameObject);
+        }
+    }
+}
