--- src/Entrega 1/Frontend/Monkey/Assets/scripts/destruirObjetosTiro.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-
-public class destruirObjetosTiro : MonoBehaviour
-{
-    private void OnTriggerEnter(Collider hit)
-    {
-        if (hit.gameObject.tag == "projetil")
-        {
-            Destroy(gameObject);
-            Destroy(hit.gameObject);
-
-        }
-    }
-}
