--- src/Entrega 1/Frontend/Monkey/Assets/scripts/destruirObjetosTiro.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 77bac1df3802a6c499173a7f425cc52a
