+++ src/Entrega 1/Frontend/Monkey/ProjectSettings/TagManager.asset
@@ -13,7 +13,8 @@ TagManager:
   - fimFase
   - textoAlvos
   - alvoBonus
-  - inimigos
+  - tiroInimigo
+  - inimigo
   layers:
   - Default
   - TransparentFX
